chore: moves jobad actions and service to career directory
