+++ src/app/career/jobads/CreateJobAdForm.tsx
@@ -1,6 +1,6 @@
 'use client'
 import styles from './CreateJobAdForm.module.scss'
-import { createJobAdAction } from '@/actions/jobAds/create'
+import { createJobAdAction } from '@/actions/career/jobAds/create'
 import TextInput from '@/components/UI/TextInput'
 import Form from '@/components/Form/Form'
 import React from 'react'
