+++ src/services/career/jobAds/ConfigVars.ts
