+++ src/services/career/jobAds/destroy.ts
